--- lib/api/mutations/create_event.rb
@@ -1,28 +1,0 @@
-require_relative 'base'
-require_relative '../../models/event'
-
-module Mutations
-  class CreateEvent < Mutations::Base
-    argument :name, String, required: true
-    argument :task_id, Integer, required: true
-
-    field :success, Boolean, null: false
-    field :errors, [String], null: false
-
-    def resolve(name:, task_id:)
-      event = Event.new(name: name, task_id: task_id)
-
-      if event.save
-        {
-          success: true,
-          errors: []
-        }
-      else
-        {
-          success: false,
-          errors: event.errors.full_messages
-        }
-      end
-    end
-  end
-end
