+++ lib/api/types/query.rb
@@ -1,10 +1,10 @@
 require_relative 'base_object'
-require_relative '../resolvers/all_events'
-require_relative '../resolvers/event'
+require_relative '../resolvers/event/all'
+require_relative '../resolvers/event/show'
 
 module Types
   class Query < Types::BaseObject
-    field :events, resolver: Resolvers::AllEvents
-    field :event, resolver: Resolvers::Event
+    field :events, resolver: Resolvers::Event::All
+    field :event, resolver: Resolvers::Event::Show
   end
 end
